Profile the effect of reindexing
Tried adding a new field for indexing to the json and tested it roughly

Before adding an index field:

    - Size of json: 125,722 bytes

    - Time elapsed from the start of navigation, average of 126 pages:
    (each page measured independent, in ms)
    {
        "responseEnd": 855.4409448818898,
        "domInteractive": 2074.834645669291,
        "domContentLoadedEventEnd": 2074.8897637795276,
        "loadEventEnd": 2283.472440944882
    }

After adding index field:

    - Size of json: 127,136 bytes

    - Time elapsed from the start of navigation, average of 126 pages:
    {
        "responseEnd": 861.1653543307086,
        "domInteractive": 2123.3700787401576,
        "domContentLoadedEventEnd": 2123.4251968503936,
        "loadEventEnd": 2335.740157480315
    }

It seems adding an index field has no benefit yet.
I'm sticking to not indexing at all.
But aside that the app has a deeper problem on performance..